+++ src/pyochain/collections/__init__.py
@@ -1,0 +1,5 @@
+"""Additional collection types."""
+
+from ._main import StableSet
+
+__all__ = ["StableSet"]